requirements.txt
psycopg2-binary==2.9.9
streamlit==1.28.0
pandas==2.1.3
numpy==1.26.2
Faker==20.1.0
ortools==9.8.3296
python-dateutil==2.9.0.post0
plotly==5.18.0